requirements.txt
alembic==1.8.1
beautifulsoup4==4.11.1
celery==5.2.7
click==8.1.3
email-validator==1.3.0
Flask==2.1.2
Flask-Admin==1.6.0
Flask-AdminLTE3==1.0.9
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
influxdb-client==1.32.0
Jinja2==3.0.3
Markdown==3.4.1
MarkupSafe==2.0.1
Pillow==9.3.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
redis==4.5.4
requests==2.28.0
rfc5424-logging-handler==1.4.3
schedule==1.2.0
shapely==2.0.1
SQLAlchemy==1.4.44
web3==6.4.0
werkzeug==2.1.2
WTForms==3.0.1
WTForms-JSON==0.3.5
xhtml2pdf==0.2.9
